requirements.txt
requests
dnspython
python-whois
colorama
rich
fpdf2
shodan
